+++ rules/style.js
@@ -98,7 +98,7 @@ module.exports = {
     // require a space before certain keywords
     'space-before-keywords': [1, 'always'],
     // require a space after certain keywords
-    'space-after-keywords': [1, 'always'],
+    'keyword-spacing': [1, 'always'],
     // require or disallow space before blocks
     'space-before-blocks': 2,
     // require or disallow space before function opening parenthesis
@@ -108,8 +108,6 @@ module.exports = {
     'space-in-parens': [2, 'never'],
     // require spaces around operators
     'space-infix-ops': 1,
-    // require a space after return, throw, and case
-    'space-return-throw-case': 2,
     // Require or disallow spaces before/after unary operators
     'space-unary-ops': 0,
     // require or disallow a space immediately following the // or /* in a comment