+++ rules/es6.js
@@ -3,7 +3,8 @@ module.exports = {
         'es6': true
     },
     parserOptions: {
-        ecmaVersion: 6
+        ecmaVersion: 6,
+        sourceType: 'module',
     },
     'rules': {
         // enforces no braces where they can be omitted